refactor: global workspace form
