+++ apps/app/components/workspace/index.ts
@@ -1,3 +1,5 @@
+export * from "./create-workspace-form";
+export * from "./delete-workspace-modal";
 export * from "./sidebar-dropdown";
 export * from "./sidebar-menu";
 export * from "./help-section";
